config: move symbol enum to cataloging pkg
